SletteModal for reiser

+++ WebApp2/ClientApp/src/app/liste/reiseListe.ts
@@ -2,7 +2,9 @@ import { Component, OnInit } from "@angular/core";
 import { HttpClient } from '@angular/common/http';
 import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
 import { Router } from '@angular/router';
+import { SletteModal } from "../modals/sletteModal";
 import { Reise } from "../Reise";
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { error } from "protractor";
 
 
@@ -14,8 +16,10 @@ import { error } from "protractor";
 export class ReiseListe {
     alleReiser: Array<Reise>;
     laster: boolean;
+  reiseTilSletting: string;
+  sletteOK: boolean;
 
-    constructor(private http: HttpClient, private router: Router) { }
+    constructor(private http: HttpClient, private router: Router, private modalService: NgbModal) { }
 
     ngOnInit() {
         this.laster = true;
@@ -33,12 +37,34 @@ export class ReiseListe {
     };
 
     sletteReise(id: number) {
+
+      this.http.get<Reise>("api/reise/" + id)
+        .subscribe(reise => {
+          this.reiseTilSletting = reise.reiseFra + " - " + reise.reiseTil;
+
+          this.visOgSlett(id);
+      },
+         error => console.log(error)
+      );
+  };
+
+  visOgSlett(id: number) {
+
+    const modal = this.modalService.open(SletteModal)
+    modal.componentInstance.sletteObjekt = this.reiseTilSletting;
+
+    modal.result.then(retur => {
+      console.log("Lukket med" + retur)
+      if (retur == "Slett") {
+
         this.http.delete("api/reise/" + id)
-            .subscribe(retur => {
-                this.hentAlleReiser();
-                this.router.navigate(['/reiseiste']);
-            },
-              error => console.log(error)
-            );
-    };
+          .subscribe(retur => {
+            this.hentAlleReiser();
+          },
+            error => console.log(error)
+          );
+      }
+      this.router.navigate(['/reiseListe']);
+    });   
+  }
 }
